backend/backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser
Pillow==11.3.0
django-filter==23.1
gunicorn==20.1.0